Handle restart from zero pods, with data
With orderly scale down we could check only the ID,
but actually index 0 might be down due to rescheduling etc
so we need to make sure there are no higher index pods.
The best we can do is to trust the service that reflects readiness.

Or we could add a service that tolerates unready endpoints, if needed.

https://mariadb.com/kb/en/library/getting-started-with-mariadb-galera-cluster/#restarting-the-cluster